requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
reportlab==4.0.4
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0